feat: Implement dashboard overview functionality with projections for revenue, category distribution, and product statistics
- Added DashboardController to handle dashboard-related requests.
- Created projections for category distribution, daily revenue, order revenue, top-rated products, and top-selling products.
- Developed response DTOs for dashboard overview, including total orders, revenue, product counts, and recent activities.
- Implemented DashboardService to aggregate data and provide an overview response.
- Updated frontend models and services to align with the new dashboard API.
- Enhanced product and stock management UI with improved button formatting and layout adjustments.

+++ activity-hub-be/src/main/java/com/softdreams/activityhub/dto/response/ProductResponse.java
@@ -19,7 +19,7 @@ public class ProductResponse {
     BigDecimal price;
     CategoryResponse category;
     double rate;
-    double totalReviews;
+    long totalReviews;
     int quantity;
     String description;
     String image;
